@@ -1,6 +1,6 @@
 module github.com/mandiant/GoReSym
 
-go 1.17
+go 1.20
 
 require (
 	golang.org/x/arch v0.0.0-20201008161808-52c3e6f60cff
